requirement.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
pprint==0.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-bitcoinrpc==1.0
requests==2.18.4
urllib3==1.22
